refactor(frontend): import baseUrl 추가 (jsconfig.json) 및 linting"
