+++ frontend/src/pages/_app.js
@@ -1,7 +1,7 @@
 import { ChakraProvider } from "@chakra-ui/react";
-import Navbar from "../components/common/Navbar";
-import { GlobalStyle } from "../styles/GlobalStyle";
-import { theme } from "../styles/theme";
+import Navbar from "components/common/Navbar";
+import { GlobalStyle } from "styles/GlobalStyle";
+import { theme } from "styles/theme";
 
 function MyApp({ Component, pageProps }) {
   return (
