+++ frontend/src/pages/index.js
@@ -1,6 +1,6 @@
-import HeroSection from "../components/pages/home/sections/HeroSection";
-import PopularSection from "../components/pages/home/sections/PopularSection";
-import RandomSection from "../components/pages/home/sections/RandomSection";
+import HeroSection from "components/pages/home/sections/HeroSection";
+import PopularSection from "components/pages/home/sections/PopularSection";
+import RandomSection from "components/pages/home/sections/RandomSection";
 
 function HomePage() {
   return (
